Extracted Printer.kt from TreePrinter.kt and moved to specific package
